ErrorHandling on repoDir not found tweaked for fetch and pull

+++ src/main/java/de/arthurpicht/meta/tasks/fetch/FetchRepoExecutor.java
@@ -5,8 +5,10 @@ import de.arthurpicht.meta.cli.executor.CommandExecutorCommons;
 import de.arthurpicht.meta.cli.output.Output;
 import de.arthurpicht.meta.cli.target.Target;
 import de.arthurpicht.meta.config.RepoConfig;
+import de.arthurpicht.meta.exception.MetaRuntimeException;
 import de.arthurpicht.meta.git.Git;
 import de.arthurpicht.meta.git.GitException;
+import de.arthurpicht.meta.helper.FilesHelper;
 import de.arthurpicht.meta.tasks.RepoExecutor;
 import de.arthurpicht.meta.tasks.TaskSummary;
 
@@ -25,20 +27,29 @@ public class FetchRepoExecutor extends RepoExecutor {
         message(repoName, "Operation pending ...");
 
         try {
+            if (!FilesHelper.isExistingDirectory(repoPath))
+                throw new MetaRuntimeException("Repo [" + repoPath.toAbsolutePath() + "] not found. Consider calling clone.");
+
             Git.fetch(repoPath);
             String currentBranch = CommandExecutorCommons.getCurrentBranch(repoConfig);
             boolean hasCommitsAhead = Git.hasCommitsAhead(repoPath, currentBranch);
             Output.deleteLastLine();
             if (!hasCommitsAhead) {
-                Output.ok(repoName,"Repo is up-to-date.");
+                Output.ok(repoName, "Repo is up-to-date.");
             } else {
                 Output.okMagenta(repoName, "Repo has commits ahead. Pull needed.");
             }
 
         } catch (GitException e) {
-            Output.error(repoName,"Git fetch failed: " + e.getMessage());
+            Output.deleteLastLine();
+            Output.error(repoName, "Git fetch failed: " + e.getMessage());
             if (ExecutionContext.isStacktrace())
                 e.printStackTrace();
+            taskSummary.addRepoFailed(repoName);
+        } catch (MetaRuntimeException e) {
+            Output.deleteLastLine();
+            Output.error(repoName, e.getMessage());
+            if (ExecutionContext.isStacktrace()) e.printStackTrace();
             taskSummary.addRepoFailed(repoName);
         }
     }
